:tada: Initial wizard attempt.
