+++ sale_order_project/wizard/sale_order_createproject_views.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<odoo>
+
+    <!-- view of the Wizard Form -->
+    <record id="sale_order_createproject_view_form" model="ir.ui.view">
+        <field name="name">sales.order.createproject.form</field>
+        <field name="model">sales.order.createproject</field>
+        <field name="arch" type="xml">
+            <form string="Create Project">
+                <group>
+                    <field name="related_project_id"/>
+                </group>
+                <footer>
+                    <button type="object" name="action_create_project" string="Create Project" class="btn-primary"/>
+                    <button string="Cancel" class="btn-default" special="cancel"/>
+                </footer>
+            </form>
+        </field>
+    </record>
+
+    <!-- action of converting, via wizard -->
+    <record id="sale_order_createproject_action" model="ir.actions.act_window">
+        <field name="name">Create Project</field>
+        <field name="res_model">sales.order.createproject</field>
+        <field name="view_type">form</field>
+        <field name="view_mode">form</field>
+        <field name="target">new</field>
+    </record>
+
+</odoo>
